Implement check funtion
