+++ src/index.js
@@ -1,10 +1,15 @@
 import './style.css';
 import {
-  mydata, display, deleteData, check, clear,
-} from './input.js';
+  mydata, display, deleteData, clear,
+} from './modules/input.js';
+
+import {
+  check, headColor,
+} from './modules/check.js';
 
 display();
 mydata();
 deleteData();
 check();
 clear();
+headColor();
